% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Inclassday2.R
\name{roll}
\alias{roll}
\title{Dice problem}
\usage{
roll(dietype = 1:6, probvalues = rep(1/length(dietype), length(dietype)))
}
\arguments{
\item{dietype}{is the amount of sides of the dice}

\item{probvalues}{the probability of getting a particular number of divots.}
}
\value{
This function returns the sum of two dice.
}
\description{
Dice problem
}
\examples{
set.seed(123)
roll()
roll(1:20)
}

