% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Inclassday2.R
\name{rp}
\alias{rp}
\title{Power function}
\usage{
rp(x, power = 1)
}
\arguments{
\item{x}{must be a numeric object}

\item{power}{must be a numeric object, raise the values in \code{power}.}
}
\value{
Raises the value x to a power
}
\description{
Power function
}
\examples{
rp(c(4:12),3)
rp(c(2,3,5),4)
rp(x=c(1,2,3), power=3)
}

