% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Inclassday2.R
\name{sq}
\alias{sq}
\title{Squaring numbers}
\usage{
sq(x)
}
\arguments{
\item{x}{is some numeric object.}
}
\value{
Squares the values in the numeric object.
}
\description{
Squaring numbers
}
\examples{
sq(c(1,2,5))
sq(3:6)
}

